file_parser/requirements.txt
flask
pyyaml
requests

